GM-37 도메인 정의
