--- src/main/java/com/gaaji/chat/domain/Message.java
@@ -1,24 +1,0 @@
-package com.gaaji.chat.domain;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-public class Message {
-    @Id
-    private String id;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User sender;
-}
