--- src/main/java/com/gaaji/chat/global/config/SpringWebSocketConfig.java
@@ -1,28 +1,0 @@
-package com.gaaji.chat.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class SpringWebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        // stomp 접속 주소 url => /ws-stomp
-        registry.addEndpoint("/ws-stomp");
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        // 메시지 구독 요청 url => 메시지를 받을 때
-        registry.enableSimpleBroker("/sub");
-
-        // 메시지 발행 요청 url => 메시지를 보낼 때
-        registry.setApplicationDestinationPrefixes("/pub");
-
-    }
-}
